+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+- Support the `@counter-style` directive.
+
 ## 0.17.3
 
 - Add markdown badges to the readme.